# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlexPos.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5088/"
    }
  ],
  "paths": {
...
    "/api/configuracion": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionNegocioDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionNegocioDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionNegocioDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Configuracion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarConfiguracionNegocioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarConfiguracionNegocioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarConfiguracionNegocioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionNegocioDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionNegocioDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionNegocioDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracion/impuestos": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Configuracion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarImpuestoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarImpuestoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarImpuestoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpuestoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracion/impuestos/{id}": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarImpuestoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarImpuestoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarImpuestoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/configuracion/impuestos/{id}/estado": {
      "patch": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecerEstadoUsuarioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecerEstadoUsuarioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecerEstadoUsuarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/configuracion/metodos-pago": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Configuracion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarMetodoPagoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarMetodoPagoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarMetodoPagoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MetodoPagoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetodoPagoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetodoPagoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuracion/metodos-pago/{id}": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarMetodoPagoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarMetodoPagoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarMetodoPagoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/configuracion/metodos-pago/{id}/estado": {
      "patch": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecerEstadoUsuarioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecerEstadoUsuarioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EstablecerEstadoUsuarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/configuracion/numeraciones-documento": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Configuracion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearNumeracionDocumentoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearNumeracionDocumentoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearNumeracionDocumentoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/configuracion/numeraciones-documento/{tipoDocumento}": {
      "put": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "tipoDocumento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPrefijoDocumentoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPrefijoDocumentoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPrefijoDocumentoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualizarPrefijoDocumentoRequest": {
        "required": [
          "prefijo"
        ],
        "type": "object",
        "properties": {
          "prefijo": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ConfiguracionNegocioDto": {
        "required": [
          "id",
          "nombreComercial",
          "razonSocial",
          "identificacionFiscal",
          "direccion",
          "telefono",
          "correo",
          "codigoMoneda",
          "permitirVentaSinStock",
          "comisionProductosHabilitada",
          "permitirSaldosPendientes",
          "facturacionHabilitada",
          "exigirEmpleadoVentaServicio",
          "fechaCreacionUtc",
          "fechaModificacionUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombreComercial": {
            "type": "string"
          },
          "razonSocial": {
            "type": [
              "null",
              "string"
            ]
          },
          "identificacionFiscal": {
            "type": [
              "null",
              "string"
            ]
          },
          "direccion": {
            "type": [
              "null",
              "string"
            ]
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "correo": {
            "type": [
              "null",
              "string"
            ]
          },
          "codigoMoneda": {
            "type": "string"
          },
          "permitirVentaSinStock": {
            "type": "boolean"
          },
          "comisionProductosHabilitada": {
            "type": "boolean"
          },
          "permitirSaldosPendientes": {
            "type": "boolean"
          },
          "facturacionHabilitada": {
            "type": "boolean"
          },
          "exigirEmpleadoVentaServicio": {
            "type": "boolean"
          },
          "fechaCreacionUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacionUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "CrearNumeracionDocumentoRequest": {
        "required": [
          "tipoDocumento",
          "prefijo",
          "siguienteNumero"
        ],
        "type": "object",
        "properties": {
          "tipoDocumento": {
            "type": [
              "null",
              "string"
            ]
          },
          "prefijo": {
            "type": [
              "null",
              "string"
            ]
          },
          "siguienteNumero": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "EstablecerEstadoUsuarioRequest": {
        "required": [
          "activo"
        ],
        "type": "object",
        "properties": {
          "activo": {
            "type": "boolean"
          }
        }
      },
...
      "GuardarConfiguracionNegocioRequest": {
        "required": [
          "nombreComercial",
          "razonSocial",
          "identificacionFiscal",
          "direccion",
          "telefono",
          "correo",
          "codigoMoneda",
          "permitirVentaSinStock",
          "comisionProductosHabilitada",
          "permitirSaldosPendientes",
          "facturacionHabilitada",
          "exigirEmpleadoVentaServicio"
        ],
        "type": "object",
        "properties": {
          "nombreComercial": {
            "type": [
              "null",
              "string"
            ]
          },
          "razonSocial": {
            "type": [
              "null",
              "string"
            ]
          },
          "identificacionFiscal": {
            "type": [
              "null",
              "string"
            ]
          },
          "direccion": {
            "type": [
              "null",
              "string"
            ]
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "correo": {
            "type": [
              "null",
              "string"
            ]
          },
          "codigoMoneda": {
            "type": [
              "null",
              "string"
            ]
          },
          "permitirVentaSinStock": {
            "type": "boolean"
          },
          "comisionProductosHabilitada": {
            "type": "boolean"
          },
          "permitirSaldosPendientes": {
            "type": "boolean"
          },
          "facturacionHabilitada": {
            "type": "boolean"
          },
          "exigirEmpleadoVentaServicio": {
            "type": "boolean"
          }
        }
      },
...
      "GuardarImpuestoRequest": {
        "required": [
          "nombre",
          "porcentaje"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "type": [
              "null",
              "string"
            ]
          },
          "porcentaje": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "GuardarMetodoPagoRequest": {
        "required": [
          "nombre",
          "requiereReferencia"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "type": [
              "null",
              "string"
            ]
          },
          "requiereReferencia": {
            "type": "boolean"
          },
          "esEfectivo": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "ImpuestoDto": {
        "required": [
          "id",
          "nombre",
          "porcentaje",
          "activo",
          "fechaCreacionUtc",
          "fechaModificacionUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "porcentaje": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "activo": {
            "type": "boolean"
          },
          "fechaCreacionUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacionUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "MetodoPagoDto": {
        "required": [
          "id",
          "nombre",
          "requiereReferencia",
          "esEfectivo",
          "activo",
          "fechaCreacionUtc",
          "fechaModificacionUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "requiereReferencia": {
            "type": "boolean"
          },
          "esEfectivo": {
            "type": "boolean"
          },
          "activo": {
            "type": "boolean"
          },
          "fechaCreacionUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacionUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}